Updated the adding car endpoint. And sending email to the user while registering and modified the data which passes in the login sending data

+++ src/main/java/com/arunpragash/car_rental/service/UserService.java
@@ -63,8 +63,8 @@ public class UserService {
         return user;
     }
 
-    public Integer getUserId(String userName) {
-        Integer id = userRepository.getIdByUserName(userName).getId();
+    public Long getUserId(String userName) {
+        Long id = userRepository.getIdByUserName(userName).getId();
         System.out.println(id);
         return id;
     }
